+++ backend/src/main/java/com/noblesi/travelplanner/admin/trip/dto/AdminTripReportDTO.java
@@ -9,8 +9,8 @@ import lombok.Getter;
 @AllArgsConstructor
 public class AdminTripReportDTO {
 
-	private int reportId;
-	private int reporterMemberId;
+	private Long reportId;
+	private Long reporterMemberId;
 	private String reporterName;
 	private String reasonCode;
 	private String reasonDetail;
